+++ UI_Escritorio/FormPersonas.cs
@@ -50,9 +50,22 @@ namespace UI_Escritorio
 
         private async void btnBorrar_Click(object sender, EventArgs e)
         {
-            int id = this.SelectedItem().IdPersona;
-            await PersonaApi.DeleteAsync(id);
-            this.GetAllAndLoad();
+             DialogResult result = MessageBox.Show("¿Está seguro que desea eliminar la persona?", "Eliminar persona", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
+            if (result == DialogResult.OK)
+            {
+                int id = this.SelectedItem().IdPersona;
+                await PersonaApi.DeleteAsync(id);
+                this.GetAllAndLoad();
+
+                MessageBox.Show("Persona eliminada con éxito", "Persona Eliminado", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                return;
+            }
+
+
+            else if (result == DialogResult.Cancel)
+            {
+                return;
+            }
         }
 
         private Persona SelectedItem()
