+++ UI_Escritorio/FormUsuario.cs
@@ -26,9 +26,23 @@ namespace UI_Escritorio
 
         private async void btnBorrar_Click(object sender, EventArgs e)
         {
-            int id = this.SelectedItem().IdUsuario;
-            await UsuariosApi.DeleteAsync(id);
-            this.GetAllAndLoad();
+
+            DialogResult result = MessageBox.Show("¿Está seguro que desea eliminar el usuario?", "Eliminar usuario", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
+            if (result == DialogResult.OK)
+            {
+                int id = this.SelectedItem().IdUsuario;
+                await UsuariosApi.DeleteAsync(id);
+                this.GetAllAndLoad();
+
+                MessageBox.Show("Usuario eliminado con éxito", "Usuario Eliminado", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                return;
+            }
+
+
+            else if (result == DialogResult.Cancel)
+            {
+                return;
+            }
         }
 
         private async void btnEditar_Click(object sender, EventArgs e)
